// packages/notation-model/src/mnx/types.ts
export type String = string;
export type Id = string;
// ...
export type Placement = 'above' | 'below' | 'auto';
// ...
export type IntegerUnsigned = number;
// ...
export type NoteValue = GlobalAttrs & {
  base: NoteValueBase;
  dots?: IntegerUnsigned;
};
export type NoteValueBase =
  | 'duplexMaxima'
  | 'maxima'
  | 'longa'
  | 'breve'
  | 'whole'
  | 'half'
  | 'quarter'
  | 'eighth'
  | '16th'
  | '32nd'
  | '64th'
  | '128th'
  | '256th'
  | '512th'
  | '1024th'
  | '2048th'
  | '4096th';
export type Time = GlobalAttrs & {
  count: PositiveInteger;
  display?: TimeSignatureDisplay;
  unit: TimeSignatureUnit;
};
export type PositiveInteger = number;
// ...
export type StaffNumber = number;
// ...
export type YesNoAuto = 'yes' | 'no' | 'auto';
// ...
export type Event = GlobalAttrs & {
  duration: NoteValue;
  fermata?: Fermata;
  kitNotes?: KitNote[];
  lyrics?: Lyrics;
  markings?: EventMarkings;
  notes?: Note[];
  rest?: Rest;
  slurs?: Slur[];
  staff?: StaffNumber;
  stemDirection?: StemDirection;
  type?: 'event';
};
// ...
export type Grace = GlobalAttrs & {
  color?: Color;
  content: Event[];
  graceType?: GraceType;
  slash?: boolean;
  type: 'grace';
};
export type GraceType = 'makeTime' | 'stealFollowing' | 'stealPrevious';
export type Tuplet = GlobalAttrs & {
  bracket?: YesNoAuto;
  content: SequenceContent;
  inner: NoteValueQuantity;
  outer: NoteValueQuantity;
  placement?: Placement;
  showNumber?: TupletDisplaySetting;
  showValue?: TupletDisplaySetting;
  staff?: StaffNumber;
  type: 'tuplet';
};
export type NoteValueQuantity = GlobalAttrs & {
  duration: NoteValue;
  multiple: PositiveInteger;
};
export type TupletDisplaySetting = 'noNumber' | 'inner' | 'both';
export type Space = GlobalAttrs & {
  duration: Fraction;
  type: 'space';
};
export type MultiNoteTremolo = GlobalAttrs & {
  content: Event[];
  individualDuration?: NoteValue;
  marks: number;
  outer: NoteValueQuantity;
  type: 'tremolo';
};
export type SequenceContent = (Event | Grace | Tuplet | Space | MultiNoteTremolo)[];
// ...
export interface GlobalAttrs {
  _c?: String;
  _x?: VendorExtensions;
  id?: Id;
}
export interface VendorExtensions {
  [k: string]: VendorDict;
}